meson: Fix git id and version string generation.
This fixes the extraction of the git id (and consequently, the version
string generation). Previously, the git id and version string, and thus
file config.h, was only regenerated on meson reconfigure. That is, when
switching branches, if no reconfiguration was triggered, the version
string was not updated. This resulted in incorrect version information,
which was especially inconvenient for dev versions, where the git commit
hash included in the version string is used to distinguish versions.